Add files via upload
I did the drivers program
I did got the input in the file line by line and I put it in the vector of the instruction (since the Dr wanted a data structure that stores the instructions)
The parsing is done in the main loop using the data structure that stores the instructions
Up until now I did LUI, AUIPC, JAL, JALR
for the memory temporarily I did it a pair of int and string (I know it is supposed to be int, int. But the string part is temporary since it helped me with tracing the PC (whether or not it is in the right place). 
The variable i in the main loop is very very important since it acts as a pointer to the instructions data structure. So we will need it when handling loops. I used it when handling JAL and JALR